Improve ansible-test env display.
